machine_learning/app/requirements.txt
# app/requirements.txt

torch==2.5.0
torchvision==0.20
basicsr @ git+https://github.com/xinntao/BasicSR.git@8d56e3a045f9fb3e1d8872f92ee4a4f07f886b0a
gfpgan==1.3.8
ultralytics==8.3.1
fastapi
uvicorn
modal
diffusers
mediapipe
loguru
Pillow
numpy
opencv-python
transformers
accelerate
safetensors
google-cloud-storage